[Fix] MDL: Silence "comparison is always false due to limited range of data type" warning. [Fix] Plugin Bridge: Silence type conversion warning in 64-bit builds.
git-svn-id: https://source.openmpt.org/svn/openmpt/trunk/OpenMPT@6638 56274372-70c3-4bfc-bfc3-4c3a0b034d27

+++ pluginBridge/Bridge.cpp
@@ -691,7 +691,7 @@ void PluginBridge::DispatchToPlugin(DispatchMsg *msg)
 	//std::cout << "about to dispatch " << msg->opcode << " to effect...";
 	//std::flush(std::cout);
 	bool exception = false;
-	msg->result = DispatchSEH(nativeEffect, msg->opcode, msg->index, static_cast<VstIntPtr>(msg->value), ptr, msg->opt, exception);
+	msg->result = static_cast<int32>(DispatchSEH(nativeEffect, msg->opcode, msg->index, static_cast<VstIntPtr>(msg->value), ptr, msg->opt, exception));
 	if(exception)
 	{
 		msg->type = MsgHeader::exceptionMsg;
